% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{cd_score}
\alias{cd_score}
\title{CD score}
\usage{
cd_score(X, X0, lambda = 1, C = NULL)
}
\arguments{
\item{X}{A numeric matrix. The training set genotypic information matrix can be given as genotype matrix (coded as -1, 0, 1) or principle component matrix (row: sample; column: marker).}

\item{X0}{A numeric matrix. The test set genotypic information matrix can be given as genotype matrix (coded as -1, 0, 1) or principle component matrix (row: sample; column: marker).}

\item{lambda}{A float number. Default lambda = 1.}

\item{C}{A numeric matrix. Default as NULL.}
}
\value{
A floating-point number, CD score.
}
\description{
This function calculate CD-score doi:10.1186/1297-9686-28-4-359 by given training set and test set. This function is imported from STPGA package developed by Deniz Akdemir \url{https://CRAN.R-project.org/package=STPGA}.
}
\examples{
data(geno)
\dontrun{cd_score(geno[1:50,], geno[51:100,])}
\dontrun{cd_score(geno[1:50,], matrix(geno[51,], nrow = 1))}

}
