% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSDFGS.R
\name{FGCM}
\alias{FGCM}
\title{Fit logistic growth curve model}
\usage{
FGCM(geno, nt = NULL, n_iter = NULL, multi.threads = TRUE)
}
\arguments{
\item{geno}{Genotype information saved as a dataframe. Columns represent variants (SNPs or PCs).}

\item{nt}{A numerical vector of training set sample size for estimating logistic growth curve parameters}

\item{n_iter}{Number of simulation of each training set size. Automatically gave a suitable number by default.}

\item{multi.threads}{Default: TRUE. Set as FALSE if you just want to run it by single thread.}
}
\value{
Estimation of parameters.
}
\description{
A function for fitting logisti growth model
}
\examples{
data(geno)
\dontrun{FGCM(geno)}
}
