% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geno}
\alias{geno}
\title{Genotype information}
\format{
A numeric matrix (PCA) with 404 rows (sample) and 404 columns (PCs).
}
\source{
\url{http://www.ricediversity.org/data/}
}
\usage{
geno
}
\description{
A PCA matrix of rice genotype information. This data was published by Zhao et al. (2011) \url{doi:10.1038/ncomms1467}
}
\examples{
data(geno)
}
\keyword{datasets}
