% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSDFGS.R
\name{nt2r}
\alias{nt2r}
\title{Simulate r-scores of each training set size}
\usage{
nt2r(geno, nt, n_iter = 30, multi.threads = TRUE)
}
\arguments{
\item{geno}{A numeric dataframe of genotype, column represent sites (genotype coding as 1, 0, -1)}

\item{nt}{Numeric. Number of training set size}

\item{n_iter}{Times of iteration. (default = 30)}

\item{multi.threads}{Default: TRUE}
}
\value{
A vector of r-scores of each iteration
}
\description{
Calculate r-scores (un-target) by in parallel.
}
\examples{
data(geno)
\dontrun{nt2r(geno, 50)}
}
