% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pev_score}
\alias{pev_score}
\title{PEV score}
\usage{
pev_score(X, X0)
}
\arguments{
\item{X}{A numeric matrix. The training set genotypic information matrix can be given as genotype matrix (coded as -1, 0, 1) or principle component matrix (row: sample; column: marker).}

\item{X0}{A numeric mareix. The test set genotypic information matrix can be given as genotype matrix (coded as -1, 0, 1) or principle component matrix (row: sample; column: marker).}
}
\value{
A floating-point number, PEV score.
}
\description{
This function calculate prediction error variance (PEV) score \url{doi:10.1186/s12711-015-0116-6} by given training set and test set.
}
\examples{
data(geno)
\dontrun{pev_score(geno[1:50, ], geno[51:100])}

}
\author{
Jen-Hsiang Ou
}
