% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r_score}
\alias{r_score}
\title{r-score}
\usage{
r_score(X, X0, lambda = 1)
}
\arguments{
\item{X}{A numeric matrix. The training set genotypic information matrix can be given as genotype matrix (coded as -1, 0, 1) or principle component matrix (row: sample; column: marker).}

\item{X0}{A numeric mareix. The test set genotypic information matrix can be given as genotype matrix (coded as -1, 0, 1) or principle component matrix (row: sample; column: marker).}

\item{lambda}{A float number. The default value is set as 1.0 and you may change it with the knowledge of heritability.}
}
\value{
A floating-point number, r-score.
}
\description{
This function calculate r-score \url{doi:10.1007/s00122-019-03387-0} by given training set and test set.
}
\examples{
data(geno)
\dontrun{r_score(geno[1:50, ], geno[51:100])}
}
\author{
Jen-Hsiang Ou
}
