% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSDFGS.R
\name{SSDFGS}
\alias{SSDFGS}
\title{Sample size determination for genomic selection}
\usage{
SSDFGS(geno, nt = NULL, n_iter = NULL, multi.threads = FALSE)
}
\arguments{
\item{geno}{A numeric data frame carried genotype information (column: PCs, row: sample)}

\item{nt}{A numeric vector carried training set sizes for r-score simulation.}

\item{n_iter}{Number of iterations for estimating parameters.}

\item{multi.threads}{Default: False. If TRUE, this function will use 75\% of threads if the computer has more than 4 threads. Multi-thread computing is only avalible in mac and linux environments.}
}
\value{
An operating curve and its information.
}
\description{
This function is designed to generate an operating curve for sample size determination
}
\examples{
data(geno)
\dontrun{SSDFGS(geno)}

}
\author{
Jen-Hsiang Ou & Po-Ya Wu
}
