% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subpop}
\alias{subpop}
\title{Sub-population information}
\format{
A character vector.
}
\source{
\url{http://www.ricediversity.org/data/}
}
\usage{
subpop
}
\description{
Sub-population information of samples. This data was published by Zhao et al. (2011) \url{doi:10.1038/ncomms1467}
}
\examples{
data(subpop)
}
\keyword{datasets}
