% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSDFGS.R
\docType{package}
\name{TSDFGS}
\alias{TSDFGS}
\title{TSDFGS}
\description{
We propose an optimality criterion to determine the required training set, r-score, which is derived directly from Pearson's correlation between the genomic estimated breeding values and phenotypic values of the test set <doi:10.1007/s00122-019-03387-0>. This package provides two main functions to determine a good training set and its size.
}
