\name{geno}
\alias{geno}
\docType{data}
\title{
Rice 44k Genomoe Data
}
\description{
A PC Matric of Rice 44k Genomoe Data
}
\usage{data("rice44kPCA")}
\format{
  A numeric matrix with 404 rows and 404 columns.
}
%\details{}
\source{
\url{http://www.ricediversity.org/data/index.cfm}
}
%\references{}
\examples{
data("rice44kPCA")
dim(geno)
}
%\keyword{datasets}
