\name{subpop}
\alias{subpop}
\docType{data}
\title{
Rice 44k Genome Data
}
\description{
Subpopulation of Each Individuals in Rice 44k Genome Data
}
\usage{data("rice44kPCA")}
\format{
  A character vector.
}
%\details{}
\source{
\url{http://www.ricediversity.org/data/index.cfm}
}
%\references{}
\examples{
data("rice44kPCA")
print(subpop)
}
%\keyword{datasets}
