% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSDFGS-package.R
\docType{package}
\name{TSDFGS-package}
\alias{TSDFGS}
\alias{TSDFGS-package}
\title{TSDFGS: Training Set Determination for Genomic Selection}
\description{
We propose an optimality criterion to determine the required training set, r-score, which is derived directly from Pearson's correlation between the genomic estimated breeding values and phenotypic values of the test set \doi{10.1007/s00122-019-03387-0}. This package provides two main functions to determine a good training set and its size.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oumarkme/TSDFGS}
  \item Report bugs at \url{https://github.com/oumarkme/TSDFGS/issues}
}

}
\author{
\strong{Maintainer}: Jen-Hsiang Ou \email{jen-hsiang.ou@imbim.uu.se} (\href{https://orcid.org/0000-0001-9305-2931}{ORCID})

Authors:
\itemize{
  \item Po-Ya Wu \email{Po-Ya.Wu@hhu.de} (\href{https://orcid.org/0000-0002-7342-2867}{ORCID})
  \item Chen-Tuo Liao \email{ctliao@ntu.edu.tw} (\href{https://orcid.org/0000-0001-9777-3701}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
